feat: Enable NVM support in zsh with detection for brew
In the zshrc file, we've enabled support for Node Version Manager (NVM). If brew command is accessible, it uses the brew installation of NVM. We've set the NVM directory and are sourcing the nvm.sh file and bash_completion file if they exist in the brew prefix nvm directory. This is to ensure that brew installation of NVM and NVM bash completion are properly loaded in the shell.